reorganize operators in `brainpy.math`
