+++ brainpy/math/event_ops.py
@@ -1,0 +1,6 @@
+
+from brainpy._src.math.operators.event_ops import (
+  event_csr_matvec,
+  event_info
+)
+
