+++ brainpy/math/jitconn_ops.py
@@ -1,0 +1,14 @@
+
+from brainpy._src.math.operators.jitconn_ops import (
+  event_matvec_prob_conn_homo_weight,
+  event_matvec_prob_conn_uniform_weight,
+  event_matvec_prob_conn_normal_weight,
+
+  matmat_prob_conn_homo_weight,
+  matmat_prob_conn_uniform_weight,
+  matmat_prob_conn_normal_weight,
+
+  matvec_prob_conn_homo_weight,
+  matvec_prob_conn_uniform_weight,
+  matvec_prob_conn_normal_weight,
+)
