+++ brainpy/math/pre_syn_post.py
@@ -1,0 +1,20 @@
+from brainpy._src.math.operators.pre_syn_post import (
+  pre2post_sum,
+  pre2post_prod,
+  pre2post_max,
+  pre2post_min,
+  pre2post_mean,
+
+  pre2post_event_sum,
+  pre2post_coo_event_sum,
+  pre2post_event_prod,
+
+  pre2syn,
+
+  syn2post_sum, syn2post,
+  syn2post_prod,
+  syn2post_max,
+  syn2post_min,
+  syn2post_mean,
+  syn2post_softmax,
+)
