+++ brainpy/math/sparse_ops.py
@@ -1,0 +1,9 @@
+from brainpy._src.math.operators.sparse_ops import (
+  cusparse_csr_matvec,
+  cusparse_coo_matvec,
+  csr_matvec,
+  sparse_matmul,
+  coo_to_csr,
+  csr_to_coo,
+  csr_to_dense
+)
