+++ docs/auto_generater.py
@@ -554,15 +554,18 @@ def generate_math_docs():
     subsections={
       'object_base': ('Basis for Object-oriented Transformations', 'brainpy.math'),
       'object_transform': ('Object-oriented Transformations', 'brainpy.math'),
-      'operators': ('Brain Dynamics Dedicated Operators', 'brainpy.math'),
-      'activations': ('Activation Functions', 'brainpy.math'),
-      'delayvars': ('Delay Variables', 'brainpy.math'),
       'environment': ('Environment Settings', 'brainpy.math'),
-      'modes': ('Computing Modes', 'brainpy.math'),
-      'arrayinterporate': ('Array Interoperability', 'brainpy.math'),
       'compat_numpy': ('Array Operators with NumPy Syntax', 'brainpy.math'),
       'compat_pytorch': ('Array Operators with PyTorch Syntax', 'brainpy.math'),
       'compat_tensorflow': ('Array Operators with TensorFlow Syntax', 'brainpy.math'),
+      'interoperability': ('Array Interoperability', 'brainpy.math'),
+      'event_ops': ('Operators for Event-driven Computation', 'brainpy.math'),
+      'jitconn_ops': ('Operators for Just-In-Time Connectivity', 'brainpy.math'),
+      'pre_syn_post': ('Operators for Pre-Syn-Post Conversion', 'brainpy.math'),
+      'sparse_ops': ('Operators for Sparse Computation', 'brainpy.math'),
+      'activations': ('Activation Functions', 'brainpy.math'),
+      'delayvars': ('Delay Variables', 'brainpy.math'),
+      'modes': ('Computing Modes', 'brainpy.math'),
       'surrogate': ('``brainpy.math.surrogate`` module: Surrogate Gradient Functions',
                     'brainpy.math.surrogate'),
       'random': ('``brainpy.math.random`` module: Random Number Generations',
